Avoid NoMethodError on very-new or broken account
When async jobs of account creation have not run yet, or have failed,
we should still (and especially) allow that account to be deleted.

The premise here is: if there is nothing to delete, don't try to delete
it.  In particular, don't break this job that needs to delete other
stuff.

+++ app/jobs/cleanup_account_job.rb
@@ -1,8 +1,9 @@
 class CleanupAccountJob < ActiveJob::Base
   def perform(account)
     account.switch do
-      solr_client.get '/solr/admin/collections', params: { action: 'DELETE', name: account.solr_endpoint.collection }
-      fcrepo_client.delete(account.fcrepo_endpoint.base_path)
+      solr_client.get '/solr/admin/collections',
+        params: { action: 'DELETE', name: account.solr_endpoint.collection } if account.solr_endpoint
+      fcrepo_client.delete(account.fcrepo_endpoint.base_path) if account.fcrepo_endpoint
     end
 
     Apartment::Tenant.drop(account.tenant)
